Split Admin Cloud components
